# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Warehouse Management API
  description: |
    API for warehouse operations, stock, approvals, reporting, and exports.

    Request and response JSON property names use snake_case. Successful responses use the common
    envelope `{success, code, message, data, errors}`. Errors use the same envelope with `success=false`,
    `data=null`, and field errors keyed by snake_case property names. Bean validation failures return 422.

    Paginated list endpoints accept optional `page` (default 1), `size` (default 10, maximum 100),
    `sort` (endpoint-supported field), `order` (`asc` or `desc`, default `desc`), and `search`.
    Authentication uses a bearer JWT. Inbound, outbound, transfer, and adjustment records follow
    draft -> submit -> approve/reject workflows; approval and rejection require their corresponding
    permission, and creators cannot review their own records. Workflow conflicts return 409 with a
    stable error code such as `*_NOT_REVIEWABLE`, `*_NOT_EDITABLE`, or `INSUFFICIENT_STOCK`; a creator
    attempting self-review receives 403 with a `*_SELF_REVIEW_FORBIDDEN` code.
  version: 1.0.0
servers:
- url: http://localhost:8081
  description: Generated server url
...
paths:
  /api/v1/warehouses/{warehouseId}/locations/{id}:
    get:
      tags:
      - Warehouse locations
      operationId: get
      parameters:
      - name: warehouseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLocationResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
    put:
      tags:
      - Warehouse locations
      operationId: update
      parameters:
      - name: warehouseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLocationResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Warehouse locations
      operationId: delete
      parameters:
      - name: warehouseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
  /api/v1/warehouses/{id}:
    get:
      tags:
      - Warehouses
      operationId: get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseWarehouseResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
    put:
      tags:
      - Warehouses
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseWarehouseResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Warehouses
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
...
  /api/v1/warehouses/{warehouseId}/locations:
    get:
      tags:
      - Warehouse locations
      operationId: list_1
      parameters:
      - name: warehouseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        description: One-based page number.
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          description: One-based page number.
          minimum: 1
      - name: size
        in: query
        description: Number of records per page (1-100).
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          description: Number of records per page (1-100).
          maximum: 100
          minimum: 1
      - name: sort
        in: query
        description: Field used to sort the results.
        required: false
        schema:
          type: string
          default: created_at
          description: Field used to sort the results.
      - name: order
        in: query
        description: "Sort direction: asc or desc."
        required: false
        schema:
          type: string
          default: desc
          description: "Sort direction: asc or desc."
          enum:
          - asc
          - desc
          pattern: (?i)^(asc|desc)$
      - name: search
        in: query
        description: Optional text filter.
        required: false
        schema:
          type: string
          description: Optional text filter.
          maxLength: 100
          minLength: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseLocationResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
    post:
      tags:
      - Warehouse locations
      operationId: create_1
      parameters:
      - name: warehouseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLocationResponse"
        "400":
          description: Malformed request or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "401":
          description: Authentication is required or the supplied token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "403":
          description: The caller lacks a required permission or warehouse access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "404":
          description: The requested resource was not found or is not visible to the
            caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "409":
          description: The request conflicts with current resource state or existing
            data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "422":
          description: Request validation or workflow rules failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
        "500":
          description: An unexpected server error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorEnvelope"
      security:
      - bearerAuth: []
...
components:
  schemas:
    LocationRequest:
      type: object
      properties:
        code:
          type: string
          maxLength: 100
          minLength: 0
        name:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
        active:
          type: boolean
      required:
      - code
      - name
    ApiResponseLocationResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/LocationResponse"
        errors:
          type: object
          additionalProperties:
            type: string
    LocationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        warehouse_id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
    WarehouseRequest:
      type: object
      properties:
        code:
          type: string
          maxLength: 100
          minLength: 0
        name:
          type: string
          maxLength: 255
          minLength: 0
        address:
          type: string
        active:
          type: boolean
      required:
      - code
      - name
    ApiResponseWarehouseResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/WarehouseResponse"
        errors:
          type: object
          additionalProperties:
            type: string
    WarehouseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        address:
          type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data: {}
        errors:
          type: object
          additionalProperties:
            type: string
...
    ApiResponsePageResponseLocationResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/PageResponseLocationResponse"
        errors:
          type: object
          additionalProperties:
            type: string
    PageResponseLocationResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/LocationResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total_elements:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
...
    ApiErrorEnvelope:
      type: object
      properties:
        success:
          default: false
        code:
          type: integer
          format: int32
        message:
          type: string
        data: {}
        errors:
          type: object
          additionalProperties:
            type: string
      required:
      - code
      - data
      - errors
      - message
      - success
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT